fix: add model summary in template
